Avoid exposing internals of service.
